+++ forth/std/prelude.py
@@ -191,6 +191,80 @@ class PreludeFunctions:
     def word():
         return Instruction(lambda state: state.word(), name="word")
 
+    # EXIT an instruction
+    @staticmethod
+    def exit():
+        return Instruction(lambda state: state.exit(), name="exit")
+
+    @staticmethod
+    def create():
+        """
+        Non standard compliant, functionality REALLY simplified
+
+        : CREATE ( -- )
+            WORD
+            state.new_word(): pop()
+        ;
+        """
+        return PreludeFunctions.word() + Instruction(
+            lambda state: state.new_word(state.drop()), name="create"
+        )
+
+    @staticmethod
+    def does():
+        return PreludeFunctions.word()
+
+    # TODO @OPTIMIZATION @WIP: Rethink this whole thing
+    @staticmethod
+    def execution_token():
+        """
+        ['] - get the 'execution token' of the following word
+
+        This just emulates the functionality, not really efficient
+        """
+        return PreludeFunctions.word() + Instruction(
+            lambda state: state.push(state.find(state.drop()))
+        )
+
+    @staticmethod
+    def immediate():
+        return Instruction(
+            lambda state: state.immediate(), name="immediate", compiled=False
+        )
+
+    @staticmethod
+    def compileto():
+        return Instruction(lambda state: state.insert(state.drop()), name="compile")
+
+    @staticmethod
+    def here():
+        return Instruction(lambda state: state.push(state.words.header))
+
+    @staticmethod
+    def return_stack_push():
+        return Instruction(lambda state: state)
+
+    @staticmethod
+    def colon():
+        """
+        DEFINE :
+            CREATE  \\ start definition using WORD
+            here    \\ current defining word (wordHeader)
+
+            >r
+
+            immediate
+
+            ['] EXIT , \\ compile EXIT to new word
+
+            r>
+
+            DOES>
+                @ execute
+        END
+        """
+        return PreludeFunctions.create() + PreludeFunctions.here()
+
     # environment
     # TODO @WIP: Use shell system to get and set env